% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasePlot_function.R
\name{gvtBasePlot}
\alias{gvtBasePlot}
\title{GVT Plot Standard}
\usage{
gvtBasePlot(g, COL)
}
\arguments{
\item{g}{GVT igraph object}

\item{COL}{Node colour - by country or industry}
}
\value{
igraph GVT plot
}
\description{
This creates a standard igraph GVT plot
}
