% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVTlayout_function.R
\name{GVTlayout}
\alias{GVTlayout}
\title{GVT layout}
\usage{
GVTlayout(g)
}
\arguments{
\item{g}{The GVT graph}
}
\value{
Tree layout matrix
}
\description{
Creates a matrix with tree layout for basic igraph plot
}
