% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVTplot_function.R
\name{GVTplot}
\alias{GVTplot}
\title{Global Value Tree Prune Plot}
\usage{
GVTplot(GVT, LAB)
}
\arguments{
\item{GVT}{The GVT network object extracted with \code{GVTprune(root)} command}

\item{LAB}{Should the plot have labels - TRUE/FALSE}
}
\value{
Global Value Tree Plot
}
\description{
This the Global Value Tree (GVT) network
}
\references{
Zhu Z, Puliga M, Cerina F, Chessa A, Riccaboni M (2015) Global Value Trees. PLoS ONE 10(5): e0126699. https://doi.org/10.1371/journal.pone.0126699
}
