% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVTprune_function.R
\name{GVTprune}
\alias{GVTprune}
\title{Global Value Tree Prune}
\usage{
GVTprune(wiot, threshold, root, layer_max)
}
\arguments{
\item{wiot}{The WIOT data loaded with \code{data(wiotyear)} command}

\item{threshold}{threshold cutoff for the GVT}

\item{root}{root node for the GVT}

\item{layer_max}{number of layers specified}
}
\value{
Global Value Tree Network Object
}
\description{
This creates a Global Value Tree (GVT) igraph network from the raw wiot data for a specific root country-industry node, with threshold and number of layers specified.
}
\references{
Zhu Z, Puliga M, Cerina F, Chessa A, Riccaboni M (2015) Global Value Trees. PLoS ONE 10(5): e0126699. https://doi.org/10.1371/journal.pone.0126699
}
