% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EdgeSwtich_function.R
\name{switchEdge}
\alias{switchEdge}
\title{Edge Switch}
\usage{
switchEdge(g)
}
\arguments{
\item{g}{graph object}
}
\value{
graph object with edges switched
}
\description{
Switches the edges so that tree layout can be calcuated
}
