% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VACel_function.R
\name{VACel}
\alias{VACel}
\title{Creating value added contribution edgelist}
\usage{
VACel(wiot)
}
\arguments{
\item{wiot}{The WIOT data loaded with \code{data(wiotyear)} command}
}
\value{
Weighted Edgelist Dataframe
#' @references Zhu Z, Puliga M, Cerina F, Chessa A, Riccaboni M (2015) Global Value Trees. PLoS ONE 10(5): e0126699. https://doi.org/10.1371/journal.pone.0126699
}
\description{
This creates a value added contribution edgelist from the raw wiot data
}
