% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VACmat_function.R
\name{VACmat}
\alias{VACmat}
\title{Creating value added contribution matrix}
\usage{
VACmat(wiot)
}
\arguments{
\item{wiot}{The WIOT data loaded with \code{data(wiotyear)} command}
}
\value{
Value added contribtuion matrix
}
\description{
This creates a value added contribution matrix from the raw wiot data
}
\references{
Zhu Z, Puliga M, Cerina F, Chessa A, Riccaboni M (2015) Global Value Trees. PLoS ONE 10(5): e0126699. https://doi.org/10.1371/journal.pone.0126699
}
