% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VACnet_function.R
\name{VACnet}
\alias{VACnet}
\title{Creating value added contribution network}
\usage{
VACnet(wiot)
}
\arguments{
\item{wiot}{The WIOT data loaded with \code{data(wiotyear)} command}
}
\value{
Weighted network - igraph object
}
\description{
This creates a value added contribution igraph network from the raw wiot data
}
\references{
Zhu Z, Puliga M, Cerina F, Chessa A, Riccaboni M (2015) Global Value Trees. PLoS ONE 10(5): e0126699. https://doi.org/10.1371/journal.pone.0126699
}
